pytest/requirements.txt
pep8
autopep8
pylint
unittest
pytest
selenium
appium-python-client
pytest-html
allure-pytest
progressbar